started rewriting for slash commands, also trying a new structure to see if this works
