+++ reqs.txt
@@ -3,7 +3,7 @@ aiosignal==1.3.1
 attrs==23.2.0
 certifi==2024.7.4
 charset-normalizer==3.3.2
-discord.py==2.4.0
+discord.py==2.5.2
 ffmpeg-python==0.2.0
 frozenlist==1.4.1
 future==1.0.0
